+++ lib/nanook/block.rb
@@ -66,13 +66,50 @@ class Nanook
       Nanook::Util.coerce_empty_string_to_type(response, Array)
     end
 
+    # Request confirmation for a block from online representative nodes.
+    # Will return immediately with a boolean to indicate if the request for
+    # confirmation was successful. Note that this boolean does not indicate
+    # the confirmation status of the block. If confirmed, your block should
+    # appear in {Nanook::Node#confirmation_history} within a short amount of
+    # time, or you can use the convenience method {Nanook::Node#confirmed_recently?}
+    #
+    # ==== Example:
+    #   block.confirm # => true
+    #
+    # @return [Boolean] if the confirmation request was sent successful
+    def confirm
+      rpc(:block_confirm, :hash)[:started] == 1
+    end
+
+    # Checks if the block appears in the list of blocks recently confirmed by
+    # online representatives (which can also be queried for using {Nanook::Node::confirmation_history}).
+    #
+    # This method can work in conjunction with {Nanook::Block#confirm},
+    # whereby you can send any block (old or new) out to online representatives to
+    # confirm. The confirmation process can take up to a couple of minutes.
+    #
+    # +false+ can indicate that the block is still in the process of being
+    # confirmed and that you should call the method again soon, or that it
+    # was confirmed earlier than the list available in {Nanook::Node::confirmation_history},
+    # or that it was not confirmed.
+    #
+    # ==== Example:
+    #   block.confirmed_recently? # => true
+    #
+    # @return [Boolean] +true+ if the block has been recently confirmed by
+    #   online representatives.
+    def confirmed_recently?
+      @rpc.call(:confirmation_history)[:confirmations].map{|h| h[:hash]}.include?(@block)
+    end
+    alias_method :recently_confirmed?, :confirmed_recently?
+
     # Generate work for a block.
     #
     # ==== Example:
     #   block.generate_work # => "2bf29ef00786a6bc"
     #
     # @param use_peers [Boolean] if set to +true+, then the node will query
-    #   its work peers (if it has any, see {Nanook::Node#work_peers}).
+    #   its work peers (if it has any, see {Nanook::WorkPeer#list}).
     #   When +false+, the node will only generate work locally (default is +false+)
     # @return [String] the work id of the work completed.
     def generate_work(use_peers: false)