+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - New `Nanook::Node#confirmation_history` method.
-- New `Nanook::Node#work_peers` method.
+- New `Nanook::Block#confirm` method.
+- New `Nanook::Block#confirmed_recently?` method.
 
 ### Changed